+++ lib/core/src/control.rs
@@ -1,2 +1,1 @@
 pub mod callback;
-pub mod event_loop;
