+++ lib/core/src/debug.rs
@@ -1,1 +1,1 @@
-pub mod stats;
+pub mod monitor;
